// types/ui-preferences.ts
export enum UiPreference {
  SUPPRESS_PII_WARNING = 'ui:suppress-pii-warning',
  SUPPRESS_REGENERATE_RESPONSE_WARNING = 'ui:suppress-regenerate-response-warning',
  SUPPRESS_GITHUB_PUSH_WARNING = 'ui:suppress-github-push-warning',
  MENUBAR_COLLAPSED = 'ui:menubar-collapsed',
  WORKFLOW_NODES_CONTAINER_COLLAPSED = 'ui:workflow-nodes-container-collapsed',
  WORKFLOWS_VIEW = 'ui:workflows-view',
  PROMPT_LIBRARY_VIEW = 'ui:prompt-library-view',
  CHAT_SIDE_PANE_WIDTH = 'ui:chat-side-pane-width',
  JOIN_USER_GROUP_DIALOG_COLLAPSED = 'ui:join-user-group-dialog-collapsed',
  START_HERE_GUIDE_SEEN = 'ui:start-here-guide-seen',
}